synthetic-data-app/gemini_client.py
import google.generativeai as genai
import os

genai.configure(api_key=os.getenv("GOOGLE_API_KEY"))

model = genai.GenerativeModel("gemini-2.0-flash")

def generate(prompt):
    response = model.generate_content(prompt)
    return response.text