requirements.txt
streamlit
psycopg2-binary
sqlparse
google-generativeai
langfuse
python-dotenv
pandas